+++ CHANGELOG.md
@@ -12,6 +12,9 @@ app.
   `user-select: all` content), avoid lingering popups, and improve performance
   (Fixes [#990](https://github.com/birchill/10ten-ja-reader/issues/990) and
   [#1033](https://github.com/birchill/10ten-ja-reader/issues/1033)).
+- (Chrome, Edge, Safari) Added support for looking up text in shadow DOM
+  elements such as are used on Bing Chat
+  [#1130](https://github.com/birchill/10ten-ja-reader/issues/1130)).
 - (Chrome, Edge, Safari) Fixed a bug where overlays would be broken on
   [MangaDex](https://mangadex.org/)
   ([#1110](https://github.com/birchill/10ten-ja-reader/issues/1110)).